requirements.txt
PyQt6
mss
google-generativeai
keyboard
pyinstaller
Pillow
